Fix/comments

+++ app/controllers/books_controller.rb
@@ -1,14 +1,41 @@
 # frozen_string_literal: true
 
 class BooksController < ApplicationController
+  include UpdateCacheConcern
+
+  before_action :authenticate_user!
+  before_action :cached_books
+  before_action :find_all_book
   before_action :find_single_book, only: %i[show edit update destroy]
 
   def index
-    @books = Book.all
+    if @cached_books.present?
+      @books = kaminari_pagination(@cached_books, 10)
+
+    else
+      @books = kaminari_pagination(@querried_all_books, 10)
+
+      Rails.cache.write('all_books_list', @querried_all_books)
+    end
+
+    @total_books = Book.count
   end
 
   def show
-    @reviews = Review.where(book_id: params[:id])
+    @cached_reviews = Rails.cache.read("all_reviews_#{params[:id]}")
+    @querried_review = @book.reviews
+
+    if @cached_reviews.present?
+      @reviews = kaminari_pagination(@cached_reviews, 10)
+
+    else
+      @reviews = kaminari_pagination(@querried_review.to_a, 10)
+      Rails.cache.write("all_reviews_#{params[:id]}", @querried_review.to_a)
+    end
+
+    add_views_cache
+
+    @total_reviews = @querried_review.count
   end
 
   def new
@@ -16,25 +43,34 @@ class BooksController < ApplicationController
   end
 
   def create
-    @book = Book.new(book_params)
-    raise ActiveRecord::RecordInvalid, @book unless @book.valid?
-    raise ActiveRecord::RecordNotSaved, @book unless @book.save
+    @book = Book.create!(book_params.merge(user_id: current_user.id))
 
-    flash[:success] = 'Book create completed'
+    update_cache(find_all_book, 'all_books_list', @querried_all_books) if @cached_books.present?
+
     redirect_to @book
   end
 
   def edit; end
 
   def update
-    raise ActiveRecord::RecordNotSaved, @book unless @book.update(book_params)
+    authorize @book, policy_class: BookPolicy
+
+    @book.update!(book_params)
+
+    update_cache(find_all_book, 'all_books_list', @querried_all_books) if @cached_books.present?
 
     flash[:success] = 'Book update completed'
+
     redirect_to @book
   end
 
   def destroy
+    authorize @book, policy_class: BookPolicy
+
     flash[:alert] = 'Failed to delete the book' unless @book.destroy
+
+    update_cache(find_all_book, 'all_books_list', @querried_all_books) if @cached_books.present?
+
     redirect_to root_path
   end
 
@@ -45,7 +81,31 @@ class BooksController < ApplicationController
   end
 
   def find_single_book
+    @cache_book = Rails.cache.read('book')
     @book = Book.find(params[:id])
-    raise ActiveRecord::RecordNotFound unless @book.present?
+  end
+
+  def find_all_book
+    @querried_all_books = Book.all.to_a
+  end
+
+  def cached_books
+    @cached_books = Rails.cache.read('all_books_list')
+  end
+
+  def kaminari_pagination(array, page)
+    Kaminari.paginate_array(array).page(params[:page]).per(page)
+  end
+
+  def add_views_cache
+    @views_cache = Rails.cache.read('views').dup || {}
+
+    if @views_cache.key?(@book.id)
+      @views_cache[@book.id] += 1
+    else
+      @views_cache[@book.id] = 1
+    end
+
+    Rails.cache.write('views', @views_cache)
   end
 end
